best model for 4

+++ DQN/her.py
@@ -40,11 +40,11 @@ def her(buffer, method, n, reward_fun):
             selected_rows = random.sample(range(n), k)
 
             selected_ind = her_ind - 1 - np.array(selected_rows)
-            selected_ind = selected_ind.astype(np.float)
+            selected_ind = selected_ind.astype(np.float32)
 
             selected_ind[selected_ind < -0.1] += buffer.replay_buffer_size
 
-            her_goals = buffer.buffer[selected_ind.astype(int), 2 * n +
+            her_goals = buffer.buffer[selected_ind.astype(np.int32), 2 * n +
                                       1:3 * n + 1]
             for goal in her_goals:
                 replay_buffer_row = np.mod(
